Refactor telnet command array in modifyDB function
This commit removes the redundant 'initSecLvl' entry from the command array in the modifyDB function. The duplicate entry was unnecessary and may have caused potential side effects during the configuration of Telnet settings.

Upgrade go-resty/resty from v2.7.0 to v2.8.0 for enhanced performance

This commit upgrades the go-resty/resty library version from 2.7.0 to 2.8.0. This was done to take advantage of the performance improvements and additional features introduced in the later version.