Refactor sendCmd to accept string arguments
Modified the sendCmd method to receive variadic string arguments, improving its usability within the loginTelnet and modifyDB functions. Previously, each command had to be manually converted to a byte array and concatenated with control characters. With the updated implementation, code readability is enhanced and redundancy is reduced.

Add reboot process in loginTelnet function

Moved the reboot process within the loginTelnet function to ensure that the device is reset before the successful Telnet message is displayed. The previous structure led to confusion as the success message was displayed before the reboot process, leading to premature signs of completion.