Improved readability and reduced redundancy in factory.go
Refactored `FactoryMode` and `Handle` functions in factory.go to improve readability and reduce redundancy. Variables `user`, `pass` and `err` are now declared once and their values updated as necessary, replacing the previous repetitious method of returning a new instance of these variables upon each error. The `tlUser`, `tlPass`, and `err` variables in the `Handle` function have also been updated to follow this new method.

This change enhances the readability of the code and prevents the unnecessary multiple declarations and returns of the same variables. It also aligns with the DRY (Don't Repeat Yourself) principle making the codebase maintainable.