Update AES test and dependencies in go.mod/go.sum
Enhanced AES test logic and updated certain dependencies.

In utils/aes_test.go, an additional test case is added to check if the decrypted string is the same as the original one, enhancing the robustness of our AES functionality tests.

In both go.mod and go.sum, the versions of `github.com/go-resty/resty/v2` and `golang.org/x/net` have been upgraded. This is to stay up to date with current dependencies and benefit from any improvements or bug fixes in these updated versions.