Bump version and enhance telnet connection handling
Updated application version and improved upon telnet connection handling for flexibility of telnet port usage.

The application version in version.go has been updated from 0.0.3 to 0.0.4 to reflect new changes added in the current push.

In app/telnet/telnet.go, the telnet connection constructor function was modified to accept an additional argument for the port number. This allows for a port number to be passed as an argument instead of hardcoded, increasing flexibility and allowing for more diverse usage of the function.

In cmd/main.go, the new argument for the telnet port number is also integrated and passed on to the telnet constructor function. This also includes a new flag for providing the telnet port from the command line.

Finally, the version of golang.org/x/net dependency has been upgraded from v0.16.0 to v0.17.0 as per indirect requirement in go.mod and go.sum, to accommodate any new features or bug fixes.