Update .goreleaser.yaml to simplify main function and cmd package
Adjusted the main function in .goreleaser.yaml to enhance usage with the cobra library, resulting in a simplified and more conventional structure. This included shifting main.go from cmd/ directory to root and integrating cobra for better CLI and flag handling. A version command was also introduced, providing users with the ability to easily access current application version.